// front/app/types/darkMode.d.ts
export type Theme = {
  deepYellow: string;
  yellow: string;
  lightYellow: string;
  brown: string;
  lightBrown: string;
  deepNavy: string;
  navy: string;
  lightNavy: string;
  deepGrey: string;
  grey: string;
  lightGrey: string;
  white: string;
};

export type DarkModeProps = {
  isDarkMode: boolean;
  theme: Theme;
};